add 00_eeg_to_bids.py
